requirements.txt
pandas==2.0.1
numpy==1.23.5
streamlit==1.21.0
plotly==5.14.1
tensorflow==2.12.0
pillow==9.5.0
scikit-learn==1.2.2